Harrison/pg vector move (#7580)
